pphysics: controller filtering changes
Controllers have a contents mask, a solid mask, and a collision group.  The solid mask is used for filtering collisions with dynamic shapes and other controllers.  The contents mask is used for filtering controller->controller collisions, the solid mask of controller A is AND'd against the contents mask of controller B.  The collision group is also optionally used for filtering controller->controller collisions.