gobj: Do CPU morphing, but GPU joint skinning
Morphing on the GPU is very complex.  Panda will perform CPU morphing if the vdata has a SliderTable, and CPU skinning if the vdata has a TransformBlendTable.  If the vdata has a TransformTable, the GPU format, it will let the shader do GPU skinning, while potentially still doing CPU morphing.