add ProxyAudioSound, for transparent async sound loads
Logic must be implemented in game code, but... the process is to start an async request for the real sound and return a ProxyAudioSound that lets the user manipulate the sound.  Then, when the real sound comes in, set the real sound on the proxy, and the proxy will copy its current state onto the real sound (such as play rate, volume, whether or not it's playing, etc).  From this point on, the user can continue using the proxy as if it were the real sound (the calls get delegated to the real sound).