fmod: implement sound occlusion
In FMOD, for Steam Audio sounds with occlusion enabled, it traces a single
ray from the sound source position to the listener position.  If the ray
is blocked, the sound's 3-band EQ factors are lowered for frequency-
dependent occlusion.  The decibel loss for each frequency band is
configurable by the user.  A generic interface is provided to allow the
audio system to trace rays using whatever underlying physics/collision
system that is desired.