map: change probe position generation
for each leaf, places a probe at N random points within empty space of the leaf.  that is, in front of all leaf boundary planes, and not inside a static prop.  this is how the probe generation in source works.  it avoids probes inside solid objects or walls that result in a black probe and weird character lighting.