optimize vertex layout for sliders/morphs
Vertex columns that get morphed are placed in a separate array to minimize the memory transfer bandwidth of the morphed vertex data to the GPU.  Static columns are placed in their own array.  Joint transform columns are lumped into the static array.  Morph deltas are also placed in another separate array as they are used on the CPU side only.