support dynamic compilation of shader object variations
Allows for much faster shader iteration times because you no longer have to recompile the entire shader object offline if you make a change to the source file.  Set shader-dynamic-compilation 1 in Config.prc, and requests to load a .sho file will redirect to the original .glsl file it was created from.  It will then compile the variations on-demand.